requirements.txt
gsheetsdb==0.1.13.1
streamlit~=1.14.1
pandas~=1.5.1
numpy~=1.23.4
google-auth~=2.14.1
chardet~=5.1.0